Fri, May 31, 2024, 2:31 PM -03:00

+++ py/week0/lesson/calculator.py
@@ -1,4 +1,4 @@
-x = int(input("X: "))
-y = int(input("Y: "))
+x = float(input("X: "))
+y = float(input("Y: "))
 
 print(x + y)
